Add experimental standard fonts helper
